+++ firmware/controllers/algo/engine.cpp
@@ -81,8 +81,8 @@ trigger_type_e getVvtTriggerType(vvt_mode_e vvtMode) {
 		return TT_ONE;
 	case VVT_2JZ:
 		return TT_VVT_JZ;
-	case VVT_MIATA_NB2:
-		return TT_VVT_MIATA_NB2;
+	case VVT_MIATA_NB:
+		return TT_VVT_MIATA_NB;
 	case VVT_BOSCH_QUICK_START:
 		return TT_VVT_BOSCH_QUICK_START;
 	case VVT_HONDA_K: